requirements.txt
pyserial==3.5
pyusb==1.2.1
